Added fold and fold_approx to bh_tree.ml.
These functions are useful for iterating over all bodies in the tree
(fold) and for iterating over bodies in the tree, using summary bodies
to approximate cells when a particular condition is met (fold_approx).
For example, the standard algorithm for computing the gravitational
force on a body using a tree can easily be written using fold_approx.